feat: per-keyword RSS feeds at /rss/<keyword>.xml
- New static endpoint /rss/<slug>.xml for each keyword on the Latest page;
  the main /rss.xml is unchanged in shape and now shares the item builder
- "rss" link in each keyword section header (Latest page only)
- Both feeds read from resolveLatest() so they agree with the page during
  the empty-month fallback
